refactor(subsonic): split route monolith into endpoint-family modules (refs #243)
Phase 3 of 4: routes/subsonic.ts (1,250 lines remaining) deleted;
handlers move verbatim into routes/subsonic/{index,browsing,search,
playlists,annotations,stream}.ts, each registering through a shared
SubsonicRouteContext (app, queries, route/publicRoute/binaryRoute).
Registration set, preHandlers, and the castToken path-gating are
unchanged; getAlbumList2's dynamic SQL stays inline in browsing.ts.
Only import-path change outside the new directory is server.ts.
