--- hub/src/routes/subsonic.ts
@@ -1,1320 +1,0 @@
-import type { FastifyPluginAsync, RouteHandlerMethod } from "fastify";
-import { Readable } from "node:stream";
-import type { Peer } from "../federation/peers.js";
-import { requireSubsonicAuth, requireSubsonicAuthBinary } from "../auth/subsonic-auth.js";
-import {
-  sendSubsonicOk,
-  sendSubsonicError,
-  sendBinaryError,
-  encodeId,
-  decodeId,
-  OPENSUBSONIC_EXTENSIONS,
-} from "./subsonic-response.js";
-import { decodeCoverArtId } from "../library/cover-art.js";
-import { isAllowedExternalArtUrl } from "./external-art.js";
-import { normalizeName } from "../library/normalize.js";
-import { SubsonicClient } from "../adapters/subsonic.js";
-import { applyTranscodeRule, buildStreamParams } from "./stream-params.js";
-import type { StreamTrackingService } from "../services/stream-tracking.js";
-import { sqliteToIso } from "../util/time.js";
-import type { ReleaseGroupRow, TrackRow } from "./subsonic/types.js";
-import {
-  contentTypeForFormat,
-  pickAlbumShareId,
-  pickArtistShareId,
-  buildSong,
-  annotateStarred,
-  annotatePlays,
-  buildAlbum,
-} from "./subsonic/builders.js";
-import { createSubsonicQueries } from "../db/queries/subsonic-queries.js";
-
-// Extend Fastify app type for stream tracking
-declare module "fastify" {
-  interface FastifyInstance {
-    streamTracking: StreamTrackingService;
-  }
-}
-// ── DB row types ──────────────────────────────────────────────────────────────
-// TrackRow / ReleaseGroupRow live in ./subsonic/types.ts (shared with builders.ts).
-
-interface ArtistRow {
-  id: string;
-  name: string;
-  albumCount: number;
-  image_url: string | null;
-}
-
-interface GenreRow {
-  genre: string;
-  albumCount: number;
-  songCount: number;
-}
-
-// ── Route plugin ──────────────────────────────────────────────────────────────
-
-export const subsonicRoutes: FastifyPluginAsync = async (app) => {
-  const preHandler = requireSubsonicAuth;
-
-  /**
-   * Register a handler for both GET and POST, with and without the .view suffix.
-   * All Subsonic endpoints accept both methods and the optional .view suffix.
-   */
-  function route(path: string, handler: RouteHandlerMethod): void {
-    app.get(path, { preHandler }, handler);
-    app.get(`${path}.view`, { preHandler }, handler);
-    app.post(path, { preHandler }, handler);
-    app.post(`${path}.view`, { preHandler }, handler);
-  }
-
-  /**
-   * Register an UNAUTHENTICATED handler (GET+POST, with/without `.view`).
-   * Only for endpoints the OpenSubsonic spec marks as callable without
-   * credentials — currently just `getOpenSubsonicExtensions`, which clients
-   * probe before login to negotiate capabilities. Do not use this for anything
-   * that reads user state or library data.
-   */
-  function publicRoute(path: string, handler: RouteHandlerMethod): void {
-    app.get(path, handler);
-    app.get(`${path}.view`, handler);
-    app.post(path, handler);
-    app.post(`${path}.view`, handler);
-  }
-
-  // ── Prepared statements (#130, factored out to a query module in #243) ──────
-  // Prepared once at plugin init, via createSubsonicQueries, rather than per-
-  // request or inline — avoids allocation churn on hot endpoints and catches
-  // SQL/schema drift at init instead of first request.
-  const queries = createSubsonicQueries(app.db);
-  const starInsertStmt = queries.starInsert;
-  const starDeleteStmt = queries.starDelete;
-  const starredArtistsStmt = queries.starredArtists;
-  const starredAlbumsStmt = queries.starredAlbums;
-  const starredSongsStmt = queries.starredSongs;
-  const starInsertTx = app.db.transaction(
-    (userId: string, rows: Array<{ kind: StarKind; raw: string }>) => {
-      for (const r of rows) starInsertStmt.run(userId, r.kind, r.raw);
-    },
-  );
-  const starDeleteTx = app.db.transaction(
-    (userId: string, rows: Array<{ kind: StarKind; raw: string }>) => {
-      for (const r of rows) starDeleteStmt.run(userId, r.kind, r.raw);
-    },
-  );
-
-  // ── ping ────────────────────────────────────────────────────────────────────
-
-  route("/ping", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-    sendSubsonicOk(reply, q, {});
-  });
-
-  // ── getLicense ──────────────────────────────────────────────────────────────
-
-  route("/getLicense", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-    sendSubsonicOk(reply, q, {
-      license: { valid: true, email: "", licenseExpires: "" },
-    });
-  });
-
-  // ── getOpenSubsonicExtensions ─────────────────────────────────────────────
-  // Capability negotiation. Per spec this endpoint is callable WITHOUT auth so
-  // clients can feature-detect before login — hence publicRoute. Returns the
-  // static OPENSUBSONIC_EXTENSIONS list (source of truth in subsonic-response).
-
-  publicRoute("/getOpenSubsonicExtensions", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-    sendSubsonicOk(reply, q, {
-      openSubsonicExtensions: OPENSUBSONIC_EXTENSIONS.map((e) => ({
-        name: e.name,
-        versions: e.versions,
-      })),
-    });
-  });
-
-  // ── getUser ─────────────────────────────────────────────────────────────────
-
-  route("/getUser", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-    const auth = request.subsonicUser!;
-    const requested = (q.username || "").trim();
-    // Subsonic spec: non-admins may only fetch their own record.
-    if (requested && requested !== auth.username && !auth.isAdmin) {
-      sendSubsonicError(reply, 50, "User is not authorized for the given operation.", q);
-      return;
-    }
-    const targetName = requested || auth.username;
-    const row = queries.userByUsername.get(targetName) as
-      | { username: string; is_admin: number }
-      | undefined;
-    if (!row) {
-      sendSubsonicError(reply, 70, "User not found.", q);
-      return;
-    }
-    const isAdmin = row.is_admin === 1;
-    sendSubsonicOk(reply, q, {
-      user: {
-        username: row.username,
-        email: "",
-        scrobblingEnabled: true,
-        adminRole: isAdmin,
-        settingsRole: isAdmin,
-        downloadRole: true,
-        uploadRole: false,
-        playlistRole: true,
-        coverArtRole: true,
-        commentRole: false,
-        podcastRole: false,
-        streamRole: true,
-        jukeboxRole: false,
-        shareRole: false,
-        videoConversionRole: false,
-      },
-    });
-  });
-
-  // ── getMusicFolders ─────────────────────────────────────────────────────────
-
-  route("/getMusicFolders", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-    // Issue #123: each known instance (local + active peers) is exposed as a
-    // MusicFolder so 3rd-party Subsonic clients can scope browsing per peer.
-    const rows = queries.musicFolders.all() as Array<{ id: number; name: string }>;
-    sendSubsonicOk(reply, q, {
-      musicFolders: { musicFolder: rows },
-    });
-  });
-
-  // ── getGenres ───────────────────────────────────────────────────────────────
-
-  route("/getGenres", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-
-    const rows = queries.genres.all() as GenreRow[];
-
-    sendSubsonicOk(reply, q, {
-      genres: {
-        genre: rows.map((r) => ({
-          value: r.genre,
-          songCount: r.songCount,
-          albumCount: r.albumCount,
-        })),
-      },
-    });
-  });
-
-  // ── getArtists ──────────────────────────────────────────────────────────────
-
-  route("/getArtists", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-
-    // INNER JOIN drops artists with no release group of their own.
-    // Featured-only / track-credit-only artists otherwise appear in client
-    // lists with zero albums and no playable content, because the artist
-    // detail view filters albums by `urg.artist_id`. Their tracks remain
-    // reachable via the album they appear on.
-    const artists = queries.artistsWithAlbumCount.all() as ArtistRow[];
-
-    const indexMap = new Map<string, typeof artists>();
-    for (const a of artists) {
-      const firstChar = a.name.trim().toUpperCase()[0] ?? "#";
-      const key = /[A-Z]/.test(firstChar) ? firstChar : "#";
-      if (!indexMap.has(key)) indexMap.set(key, []);
-      indexMap.get(key)!.push(a);
-    }
-
-    const index = [...indexMap.entries()]
-      .sort(([a], [b]) => (a === "#" ? 1 : b === "#" ? -1 : a.localeCompare(b)))
-      .map(([name, list]) => ({
-        name,
-        artist: list.map((a) => ({
-          id: encodeId("ar", a.id),
-          name: a.name,
-          albumCount: a.albumCount,
-          coverArt: a.image_url ?? undefined,
-        })),
-      }));
-
-    sendSubsonicOk(reply, q, {
-      artists: { ignoredArticles: "The An A", index },
-    });
-  });
-
-  // ── getIndexes ──────────────────────────────────────────────────────────────
-
-  route("/getIndexes", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-
-    // INNER JOIN: see /getArtists above for the rationale.
-    const artists = queries.artistIndexRows.all() as ArtistRow[];
-
-    const indexMap = new Map<string, typeof artists>();
-    for (const a of artists) {
-      const firstChar = a.name.trim().toUpperCase()[0] ?? "#";
-      const key = /[A-Z]/.test(firstChar) ? firstChar : "#";
-      if (!indexMap.has(key)) indexMap.set(key, []);
-      indexMap.get(key)!.push(a);
-    }
-
-    const index = [...indexMap.entries()]
-      .sort(([a], [b]) => (a === "#" ? 1 : b === "#" ? -1 : a.localeCompare(b)))
-      .map(([name, list]) => ({
-        name,
-        artist: list.map((a) => ({
-          id: encodeId("ar", a.id),
-          name: a.name,
-          albumCount: a.albumCount,
-        })),
-      }));
-
-    sendSubsonicOk(reply, q, {
-      indexes: {
-        ignoredArticles: "The An A",
-        lastModified: Date.now(),
-        index,
-      },
-    });
-  });
-
-  // ── getArtist ───────────────────────────────────────────────────────────────
-
-  route("/getArtist", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-
-    let artistId: string;
-    try {
-      artistId = decodeId(q.id ?? "", "ar");
-    } catch {
-      sendSubsonicError(reply, 70, "Artist not found", q);
-      return;
-    }
-
-    const artist = queries.artistById.get(artistId) as
-      | { id: string; name: string; image_url: string | null }
-      | undefined;
-
-    if (!artist) {
-      sendSubsonicError(reply, 70, "Artist not found", q);
-      return;
-    }
-
-    const albums = queries.albumsForArtist.all(artistId) as ReleaseGroupRow[];
-
-    const shareId = pickArtistShareId(app.db, artist.id);
-
-    const builtAlbums = albums.map(buildAlbum);
-    annotateStarred(app.db, request.subsonicUser?.id, "album", "al", builtAlbums);
-    annotatePlays(app.playEvents, request.subsonicUser?.id, "album", "al", builtAlbums);
-    const artistObj: {
-      id: string;
-      name: string;
-      albumCount: number;
-      coverArt?: string;
-      shareId?: string;
-      album: ReturnType<typeof buildAlbum>[];
-      starred?: string;
-    } = {
-      id: encodeId("ar", artist.id),
-      name: artist.name,
-      albumCount: albums.length,
-      coverArt: artist.image_url ?? undefined,
-      shareId: shareId ?? undefined,
-      album: builtAlbums,
-    };
-    annotateStarred(app.db, request.subsonicUser?.id, "artist", "ar", [artistObj]);
-    sendSubsonicOk(reply, q, { artist: artistObj });
-  });
-
-// ── getArtistInfo2 ────────────────────────────────────────────────────────
-
-  route("/getArtistInfo2", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-
-    let artistId: string;
-    try {
-      artistId = decodeId(q.id ?? "", "ar");
-    } catch {
-      sendSubsonicError(reply, 70, "Artist not found", q);
-      return;
-    }
-
-    const artistRow = queries.artistInfoById.get(artistId) as
-      | { id: string; name: string; musicbrainz_id: string | null; image_url: string | null }
-      | undefined;
-
-    if (!artistRow) {
-      sendSubsonicError(reply, 70, "Artist not found", q);
-      return;
-    }
-
-    // Get image URL from unified_artists (may be Last.fm URL or encoded cover art ID)
-    let imageUrl: string | undefined;
-    if (artistRow.image_url) {
-      if (artistRow.image_url.startsWith("https://")) {
-        // It's a Last.fm URL, return directly
-        imageUrl = artistRow.image_url;
-      } else {
-        // It's an encoded cover art ID, return as-is for client to resolve
-        imageUrl = artistRow.image_url;
-      }
-    }
-
-    // If no image URL and Last.fm is enabled, try to fetch from Last.fm
-    if (!imageUrl && app.lastFmClient?.isEnabled()) {
-      try {
-        const lastFmInfo = await app.lastFmClient.getArtistInfo(
-          artistRow.name,
-          artistRow.musicbrainz_id ?? undefined
-        );
-
-        if (lastFmInfo) {
-          const bestImage = app.lastFmClient.getBestImage(lastFmInfo);
-          if (bestImage) {
-            // Cache the image URL in the database
-            queries.updateArtistImageUrl.run(bestImage, artistId);
-            imageUrl = bestImage;
-            request.log.info(`Cached Last.fm image for artist ${artistRow.name}`);
-          }
-        }
-      } catch (err) {
-        request.log.warn(`Failed to fetch Last.fm info for artist ${artistRow.name}: ${err}`);
-      }
-    }
-
-    sendSubsonicOk(reply, q, {
-      artistInfo2: {
-        artist: {
-          id: encodeId("ar", artistRow.id),
-          name: artistRow.name,
-          musicBrainzId: artistRow.musicbrainz_id ?? undefined,
-        },
-        smallImageUrl: imageUrl,
-        mediumImageUrl: imageUrl,
-        largeImageUrl: imageUrl,
-        musicBrainzId: artistRow.musicbrainz_id ?? undefined,
-      },
-    });
-  });
-
-  // ── getAlbumList2 ───────────────────────────────────────────────────────────
-
-  route("/getAlbumList2", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-    const type = q.type ?? "newest";
-    const size = Math.min(parseInt(q.size ?? "10", 10), 500);
-    const offset = parseInt(q.offset ?? "0", 10);
-    const fromYear = q.fromYear ? parseInt(q.fromYear, 10) : undefined;
-    const toYear = q.toYear ? parseInt(q.toYear, 10) : undefined;
-    const genre = q.genre;
-    // Standard Subsonic param (issue #123). q.instanceId is an EOL alias
-    // kept for in-tree callers mid-migration — slated for removal; do not
-    // adopt in new code. See docs/opensubsonic.md.
-    let instanceId: string | undefined = q.instanceId;
-    if (!instanceId && q.musicFolderId) {
-      const row = queries.instanceByMusicFolderId.get(
-        parseInt(q.musicFolderId, 10),
-      ) as { id: string } | undefined;
-      // Unknown folder id → empty result, matching how Subsonic clients expect
-      // an unrecognized scope to surface (no rows rather than an error).
-      if (!row) {
-        return sendSubsonicOk(reply, q, { albumList2: { album: [] } });
-      }
-      instanceId = row.id;
-    }
-
-    let orderBy = "urg.created_at DESC";
-    let where = "WHERE 1=1";
-    const params: unknown[] = [];
-
-    // #197: type=frequent / type=recent rank by the requesting user's play
-    // history. A LEFT JOIN onto per-release-group aggregates lets us order by
-    // play count / last-played and (for these two types) exclude never-played
-    // albums — matching how Subsonic clients expect "frequent"/"recent". The
-    // join's `?` (user id) is the first bound param, so it's prepended below.
-    const usePlayJoin = type === "frequent" || type === "recent";
-    const playJoin = usePlayJoin
-      ? `LEFT JOIN (
-           SELECT ur2.release_group_id AS rg,
-                  COUNT(*) AS play_count,
-                  MAX(pe.played_at) AS last_played
-           FROM play_events pe
-           JOIN unified_tracks ut2 ON ut2.id = pe.unified_track_id
-           JOIN unified_releases ur2 ON ur2.id = ut2.release_id
-           WHERE pe.user_id = ?
-           GROUP BY ur2.release_group_id
-         ) pc ON pc.rg = urg.id`
-      : "";
-    const joinParams: unknown[] = usePlayJoin ? [request.subsonicUser.id] : [];
-    // Reuse the playJoin's aggregate for playCount/played instead of a second
-    // pass via getAlbumStats (#197). Functionally dependent on urg.id (one pc
-    // row per release group), so safe under GROUP BY urg.id.
-    const playCols = usePlayJoin ? ", pc.play_count, pc.last_played" : "";
-
-    // type=starred — restrict to albums starred by the requesting user. (#104)
-    if (type === "starred") {
-      where +=
-        " AND EXISTS (SELECT 1 FROM user_stars us " +
-        "WHERE us.user_id = ? AND us.kind = 'album' AND us.target_id = urg.id)";
-      params.push(request.subsonicUser.id);
-    }
-
-    // EXISTS avoids row multiplication when an album has multiple sources.
-    if (instanceId) {
-      where +=
-        " AND EXISTS (SELECT 1 FROM unified_releases ur2 " +
-        "JOIN unified_release_sources urs ON urs.unified_release_id = ur2.id " +
-        "JOIN instance_albums ia ON ia.id = urs.instance_album_id " +
-        "WHERE ur2.release_group_id = urg.id AND ia.instance_id = ?)";
-      params.push(instanceId);
-    }
-
-    switch (type) {
-      case "alphabeticalByName":
-        orderBy = "urg.name_normalized ASC";
-        break;
-      case "alphabeticalByArtist":
-        orderBy = "ua.name_normalized ASC, urg.name_normalized ASC";
-        break;
-      case "byYear":
-        if (fromYear !== undefined) {
-          where += " AND urg.year >= ?";
-          params.push(fromYear);
-        }
-        if (toYear !== undefined) {
-          where += " AND urg.year <= ?";
-          params.push(toYear);
-        }
-        orderBy =
-          (fromYear ?? 0) <= (toYear ?? 9999)
-            ? "urg.year ASC"
-            : "urg.year DESC";
-        break;
-      case "byGenre":
-        if (genre) {
-          where += " AND urg.genre = ?";
-          params.push(genre);
-        }
-        orderBy = "urg.name_normalized ASC";
-        break;
-      case "random":
-        orderBy = "RANDOM()";
-        break;
-      case "starred":
-        orderBy = "urg.name_normalized ASC";
-        break;
-      case "frequent":
-        // Most-played first; restrict to albums the user has actually played.
-        // urg.id breaks ties so LIMIT/OFFSET paging is stable across windows.
-        where += " AND pc.play_count IS NOT NULL";
-        orderBy = "pc.play_count DESC, pc.last_played DESC, urg.id";
-        break;
-      case "recent":
-        // Most-recently-played first; played albums only. SQLite sorts NULLs
-        // last under DESC, but the WHERE already excludes the never-played.
-        // urg.id breaks ties so LIMIT/OFFSET paging is stable across windows.
-        where += " AND pc.play_count IS NOT NULL";
-        orderBy = "pc.last_played DESC, urg.id";
-        break;
-      // highest (ratings) — not tracked; fall back to newest.
-      default:
-        orderBy = "urg.created_at DESC";
-        break;
-    }
-
-    params.push(size, offset);
-
-    const albums = app.db
-      .prepare(
-        `SELECT urg.id, urg.name, urg.artist_id, ua.name AS artist_name,
-          urg.year, urg.genre, urg.image_url, urg.created_at,
-          COUNT(ut.id) AS songCount${playCols}
-        FROM unified_release_groups urg
-        JOIN unified_artists ua ON ua.id = urg.artist_id
-        ${playJoin}
-        LEFT JOIN unified_releases ur ON ur.release_group_id = urg.id
-        LEFT JOIN unified_tracks ut ON ut.release_id = ur.id
-        ${where}
-        GROUP BY urg.id
-        ORDER BY ${orderBy}
-        LIMIT ? OFFSET ?`,
-      )
-      .all(...joinParams, ...params) as ReleaseGroupRow[];
-
-    const builtAlbums = albums.map(buildAlbum);
-    annotateStarred(app.db, request.subsonicUser?.id, "album", "al", builtAlbums);
-    if (usePlayJoin) {
-      // frequent/recent: the playJoin already computed each album's play
-      // aggregate for this page — reuse it instead of re-querying getAlbumStats
-      // (#197). Same per-user, source-agnostic count, same ISO conversion.
-      for (let i = 0; i < albums.length; i++) {
-        const count = albums[i].play_count;
-        if (count != null && count > 0) {
-          const a = builtAlbums[i] as (typeof builtAlbums)[number] & {
-            playCount?: number;
-            played?: string;
-          };
-          a.playCount = count;
-          if (albums[i].last_played) a.played = sqliteToIso(albums[i].last_played!);
-        }
-      }
-    } else {
-      annotatePlays(app.playEvents, request.subsonicUser?.id, "album", "al", builtAlbums);
-    }
-    sendSubsonicOk(reply, q, { albumList2: { album: builtAlbums } });
-  });
-
-  // ── getAlbum ────────────────────────────────────────────────────────────────
-
-  route("/getAlbum", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-
-    let rgId: string;
-    try {
-      rgId = decodeId(q.id ?? "", "al");
-    } catch {
-      sendSubsonicError(reply, 70, "Album not found", q);
-      return;
-    }
-
-    const rg = queries.releaseGroupById.get(rgId) as
-      | {
-          id: string;
-          name: string;
-          artist_id: string;
-          artist_name: string;
-          year: number | null;
-          genre: string | null;
-          image_url: string | null;
-          created_at: string | null;
-        }
-      | undefined;
-
-    if (!rg) {
-      sendSubsonicError(reply, 70, "Album not found", q);
-      return;
-    }
-
-    // Pick the release with the most tracks (fall back to first by id)
-    const release = queries.bestReleaseForGroup.get(rgId) as { id: string } | undefined;
-
-    const tracks: TrackRow[] = release
-      ? (queries.tracksForRelease.all(release.id) as TrackRow[])
-      : [];
-
-    const totalDuration = tracks.reduce(
-      (sum, t) => sum + (t.duration_ms != null ? Math.round(t.duration_ms / 1000) : 0),
-      0,
-    );
-
-    const shareId = pickAlbumShareId(app.db, rg.id);
-
-    const builtSongs = tracks.map(buildSong);
-    annotateStarred(app.db, request.subsonicUser?.id, "track", "t", builtSongs);
-    annotatePlays(app.playEvents, request.subsonicUser?.id, "track", "t", builtSongs);
-    const albumObj: {
-      id: string;
-      name: string;
-      artist: string;
-      artistId: string;
-      coverArt?: string;
-      songCount: number;
-      duration: number;
-      year?: number;
-      genre?: string;
-      shareId?: string;
-      song: ReturnType<typeof buildSong>[];
-      starred?: string;
-      created?: string;
-      playCount?: number;
-      played?: string;
-    } = {
-      id: encodeId("al", rg.id),
-      name: rg.name,
-      artist: rg.artist_name,
-      artistId: encodeId("ar", rg.artist_id),
-      coverArt: rg.image_url ?? undefined,
-      songCount: tracks.length,
-      duration: totalDuration,
-      year: rg.year ?? undefined,
-      genre: rg.genre ?? undefined,
-      shareId: shareId ?? undefined,
-      song: builtSongs,
-      created: rg.created_at ? sqliteToIso(rg.created_at) : undefined,
-    };
-    annotateStarred(app.db, request.subsonicUser?.id, "album", "al", [albumObj]);
-    annotatePlays(app.playEvents, request.subsonicUser?.id, "album", "al", [albumObj]);
-    sendSubsonicOk(reply, q, {
-      album: albumObj,
-    });
-  });
-
-  // ── getSong ─────────────────────────────────────────────────────────────────
-
-  route("/getSong", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-
-    let trackId: string;
-    try {
-      trackId = decodeId(q.id ?? "", "t");
-    } catch {
-      sendSubsonicError(reply, 70, "Song not found", q);
-      return;
-    }
-
-    const row = queries.trackForSong.get(trackId) as TrackRow | undefined;
-
-    if (!row) {
-      sendSubsonicError(reply, 70, "Song not found", q);
-      return;
-    }
-
-    const built = buildSong(row);
-    annotateStarred(app.db, request.subsonicUser?.id, "track", "t", [built]);
-    annotatePlays(app.playEvents, request.subsonicUser?.id, "track", "t", [built]);
-    sendSubsonicOk(reply, q, { song: built });
-  });
-
-  // ── search3 ─────────────────────────────────────────────────────────────────
-
-  route("/search3", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-    const query = q.query ?? "";
-    const artistCount = parseInt(q.artistCount ?? "20", 10);
-    const albumCount = parseInt(q.albumCount ?? "20", 10);
-    const songCount = parseInt(q.songCount ?? "20", 10);
-    const artistOffset = parseInt(q.artistOffset ?? "0", 10);
-    const albumOffset = parseInt(q.albumOffset ?? "0", 10);
-    const songOffset = parseInt(q.songOffset ?? "0", 10);
-    const like = `%${normalizeName(query)}%`;
-
-    // ID lookup: allow pasting an internal ID (optionally prefixed ar/al/t)
-    // or a MusicBrainz ID into the search box. Strip known prefixes so the
-    // bare UUID matches id/musicbrainz_id columns directly.
-    const trimmed = query.trim();
-    const artistIdCandidate = trimmed.startsWith("ar") ? trimmed.slice(2) : trimmed;
-    const albumIdCandidate = trimmed.startsWith("al") ? trimmed.slice(2) : trimmed;
-    const songIdCandidate = trimmed.startsWith("t") ? trimmed.slice(1) : trimmed;
-
-    const artists = queries.searchArtists.all(
-      like,
-      trimmed,
-      artistIdCandidate,
-      trimmed,
-      artistIdCandidate,
-      trimmed,
-      artistCount,
-      artistOffset,
-    ) as ArtistRow[];
-
-    const albums = queries.searchAlbums.all(
-      like,
-      trimmed,
-      albumIdCandidate,
-      trimmed,
-      albumIdCandidate,
-      trimmed,
-      albumCount,
-      albumOffset,
-    ) as ReleaseGroupRow[];
-
-    const songs = queries.searchSongs.all(
-      like,
-      trimmed,
-      songIdCandidate,
-      trimmed,
-      songIdCandidate,
-      trimmed,
-      songCount,
-      songOffset,
-    ) as TrackRow[];
-
-    const builtArtists = artists.map((a) => ({
-      id: encodeId("ar", a.id),
-      name: a.name,
-      albumCount: a.albumCount,
-      coverArt: a.image_url ?? undefined,
-    }));
-    const builtAlbums = albums.map(buildAlbum);
-    const builtSongs = songs.map(buildSong);
-    annotateStarred(app.db, request.subsonicUser?.id, "artist", "ar", builtArtists);
-    annotateStarred(app.db, request.subsonicUser?.id, "album", "al", builtAlbums);
-    annotateStarred(app.db, request.subsonicUser?.id, "track", "t", builtSongs);
-    annotatePlays(app.playEvents, request.subsonicUser?.id, "album", "al", builtAlbums);
-    annotatePlays(app.playEvents, request.subsonicUser?.id, "track", "t", builtSongs);
-    sendSubsonicOk(reply, q, {
-      searchResult3: {
-        artist: builtArtists,
-        album: builtAlbums,
-        song: builtSongs,
-      },
-    });
-  });
-
-  // ── getCoverArt ─────────────────────────────────────────────────────────────
-  // Binary endpoint: uses requireSubsonicAuthBinary so auth failures return
-  // real HTTP error codes instead of a 200+JSON Subsonic envelope.
-
-  function binaryRoute(path: string, handler: RouteHandlerMethod): void {
-    const binaryPreHandler = requireSubsonicAuthBinary;
-    app.get(path, { preHandler: binaryPreHandler }, handler);
-    app.get(`${path}.view`, { preHandler: binaryPreHandler }, handler);
-    app.post(path, { preHandler: binaryPreHandler }, handler);
-    app.post(`${path}.view`, { preHandler: binaryPreHandler }, handler);
-  }
-
-  binaryRoute("/getCoverArt", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-    const id = q.id ?? "";
-    const size = q.size;
-
-    // Raw external URL `id` (e.g. 3rd-party Subsonic clients echoing back the
-    // `coverArt` value we returned for an artist image). Skip decodeCoverArtId
-    // — the embedded `https:` colon would mis-parse as an instance prefix —
-    // and route through the external-fetch branch below.
-    let instanceId: string;
-    let coverArtId: string;
-    if (id.startsWith("http://") || id.startsWith("https://")) {
-      instanceId = "local";
-      coverArtId = id;
-    } else {
-      try {
-        const decoded = decodeCoverArtId(id);
-        instanceId = decoded.instanceId;
-        coverArtId = decoded.coverArtId;
-      } catch {
-        sendBinaryError(reply, 400, "Invalid cover art ID");
-        return;
-      }
-    }
-
-    const cacheKey = size ? `${id}:${size}` : id;
-
-    // Check cache first (covers both local and peer art). Skip a hit if the
-    // cached entry is not a real image — old Navidrome "missing cover" XML
-    // envelopes may have been cached before the upstream-validation fix.
-    const cached = app.artCache.get(cacheKey);
-    if (cached && cached.contentType.startsWith("image/") && cached.data.length > 0) {
-      reply.raw.writeHead(200, {
-        "content-type": cached.contentType,
-        "content-length": String(cached.data.length),
-        "cache-control": "public, max-age=2592000",
-        "x-cache": "HIT",
-      });
-      reply.raw.end(cached.data);
-      return;
-    }
-    if (cached) {
-      // Poisoned entry — drop it and re-fetch.
-      app.artCache.delete(cacheKey);
-    }
-
-    let response: Response;
-
-    // External URL (e.g. fanart.tv album cover stored when Navidrome had none).
-    // The decoded coverArtId is the full https:// URL — fetch it directly
-    // instead of routing through Navidrome or a peer proxy. Restricted to a
-    // fanart.tv hostname allowlist; peer-supplied URLs to other hosts are
-    // rejected to prevent SSRF.
-    if (coverArtId.startsWith("http://") || coverArtId.startsWith("https://")) {
-      if (!isAllowedExternalArtUrl(coverArtId)) {
-        sendBinaryError(reply, 400, "Disallowed external art URL");
-        return;
-      }
-      try {
-        response = await fetch(coverArtId);
-      } catch {
-        sendBinaryError(reply, 502, "Failed to fetch external art");
-        return;
-      }
-    } else if (instanceId !== "local") {
-      // Peer art routing via /proxy/rest/getCoverArt — Ed25519-signed request to the peer's proxy.
-      // The signing path must include the /proxy prefix (as seen by the peer's Fastify router).
-      const peer = app.peerRegistry.peers.get(instanceId);
-      if (!peer) {
-        sendBinaryError(reply, 404, "Peer not found");
-        return;
-      }
-      try {
-        const artParams = new URLSearchParams({ id: coverArtId });
-        if (size) artParams.set("size", size);
-        const signingPath = `/proxy/rest/getCoverArt?${artParams.toString()}`;
-        // Substitute peer.proxyUrl as the base so the HTTP request goes to the correct host.
-        const proxyPeer: Peer = { ...peer, url: peer.proxyUrl };
-        response = await app.federatedFetch(
-          proxyPeer,
-          signingPath,
-          { asUser: request.subsonicUser.username },
-        );
-      } catch {
-        sendBinaryError(reply, 502, "Failed to fetch art from peer");
-        return;
-      }
-    } else {
-      // Local Navidrome art.
-      // TODO(phase-5): route through /proxy/rest/getCoverArt (internal inject) once local
-      // reads are uniformly proxied. SubsonicClient hits Navidrome directly for now.
-      const client = new SubsonicClient({
-        url: app.config.navidromeUrl,
-        username: app.config.navidromeUsername,
-        password: app.config.navidromePassword,
-      });
-      try {
-        const sizeNum = size ? parseInt(size, 10) : undefined;
-        response = await client.getCoverArt(coverArtId, sizeNum);
-      } catch {
-        sendBinaryError(reply, 502, "Failed to fetch art from Navidrome");
-        return;
-      }
-    }
-
-    if (!response.ok) {
-      sendBinaryError(reply, response.status === 404 ? 404 : 502, "Art not found");
-      return;
-    }
-
-    if (!response.body) {
-      sendBinaryError(reply, 502, "Empty response from upstream");
-      return;
-    }
-
-    const chunks: Uint8Array[] = [];
-    const reader = response.body.getReader();
-    while (true) {
-      const { done, value } = await reader.read();
-      if (done) break;
-      if (value) chunks.push(value);
-    }
-    const buffer = Buffer.concat(chunks);
-    const contentType = response.headers.get("content-type") || "image/jpeg";
-
-    // Navidrome answers a missing cover with HTTP 200 + a Subsonic XML/JSON
-    // error envelope (Content-Type: text/xml or application/json), not an
-    // image. Don't cache or serve those — let the client treat it as 404 so
-    // a future re-sync can supply a real cover (e.g. via fanart.tv).
-    if (buffer.length === 0 || !contentType.startsWith("image/")) {
-      sendBinaryError(reply, 404, "Art not found");
-      return;
-    }
-
-    app.artCache.put(cacheKey, buffer, contentType);
-
-    reply.raw.writeHead(200, {
-      "content-type": contentType,
-      "content-length": String(buffer.length),
-      "cache-control": "public, max-age=2592000",
-      "x-cache": "MISS",
-    });
-    reply.raw.end(buffer);
-  });
-
-  // ── stream ──────────────────────────────────────────────────────────────────
-
-async function handleStream(request: Parameters<RouteHandlerMethod>[0], reply: Parameters<RouteHandlerMethod>[1]) {
-  const q = request.query as Record<string, string>;
-
-let trackId: string;
-try {
-  trackId = decodeId(q.id ?? "", "t");
-  request.log.info(`Stream request: decoded trackId = ${trackId}`);
-} catch {
-  request.log.warn(`Stream request: failed to decode track ID from ${q.id}`);
-  sendBinaryError(reply, 400, "Invalid track ID");
-  return;
-}
-
-  // Get track info for streaming
-  const trackRow = queries.trackForStream.get(trackId) as
-    | { id: string; title: string; artist_name: string; duration_ms: number | null }
-    | undefined;
-
-  if (!trackRow) {
-    request.log.warn(`Stream tracking: track ${trackId} not found in unified_tracks`);
-  }
-
-  // Defer stream tracking start until after source/transcode resolution
-  // so we can record format/bitrate/source/transcode flags up front.
-  let streamOpId: string | undefined;
-
-  // Source selection happens at merge time (merge.ts sets preferred = 1).
-  // At stream time we just look up THE source for this unified track.
-  const best = queries.preferredSourceForStream.get(trackId) as
-    | { instance_id: string; format: string | null; bitrate: number | null; remote_id: string }
-    | undefined;
-
-  if (!best) {
-    sendBinaryError(reply, 404, "Track not found");
-    return;
-  }
-
-  const streamParams = applyTranscodeRule(buildStreamParams(q), {
-    format: best.format,
-    bitrate: best.bitrate,
-  });
-
-  // Forward Range only for raw passthrough streams. If transcoding is in
-  // play, byte offsets in the upstream response would not map to the
-  // transcoded bytes the caller expects. (#97 — transcoded seek tracked
-  // separately.)
-  const cap = Number(streamParams.get("maxBitRate")) || Infinity;
-  const srcBr = best.bitrate ?? Infinity;
-  const isRaw =
-    !streamParams.has("format") &&
-    !streamParams.has("timeOffset") &&
-    srcBr <= cap;
-  const rangeHeader =
-    isRaw && typeof request.headers.range === "string"
-      ? request.headers.range
-      : undefined;
-
-  if (trackRow) {
-    const transcoded = streamParams.has("format") || (Number.isFinite(cap) && srcBr > cap);
-    // Cast-token-authed stream → tag activity with the appropriate kind
-    // (#218). `dlna=1` flag set by buildStreamUrl for DLNA renderers; cast
-    // token without the flag is from a Sonos device.
-    const streamKind: "subsonic" | "cast" | "dlna" = q.castToken
-      ? q.dlna === "1"
-        ? "dlna"
-        : "cast"
-      : "subsonic";
-    streamOpId = app.streamTracking.start({
-      kind: streamKind,
-      username: request.subsonicUser.username,
-      trackId: trackRow.id,
-      trackTitle: trackRow.title,
-      artistName: trackRow.artist_name,
-      clientName: q.c ?? null,
-      clientVersion: q.v ?? null,
-      sourceKind: best.instance_id === "local" ? "local" : "peer",
-      sourcePeerId: best.instance_id === "local" ? null : best.instance_id,
-      format: best.format,
-      bitrate: best.bitrate,
-      transcoded,
-      maxBitrate: Number.isFinite(cap) ? cap : null,
-    });
-  }
-
-  let response: Response;
-  let bytesTransferred = 0;
-
-  if (best.instance_id === "local") {
-    const client = new SubsonicClient({
-      url: app.config.navidromeUrl,
-      username: app.config.navidromeUsername,
-      password: app.config.navidromePassword,
-    });
-    try {
-      const opts: { format?: string; maxBitRate?: number; timeOffset?: number; estimateContentLength?: boolean; range?: string } = {};
-      const fmt = streamParams.get("format");
-      const br = streamParams.get("maxBitRate");
-      const to = streamParams.get("timeOffset");
-      const ecl = streamParams.get("estimateContentLength");
-      if (fmt) opts.format = fmt;
-      if (br) opts.maxBitRate = parseInt(br, 10);
-      if (to) opts.timeOffset = parseInt(to, 10);
-      if (ecl === "true") opts.estimateContentLength = true;
-      if (rangeHeader) opts.range = rangeHeader;
-      response = await client.stream(best.remote_id, opts);
-    } catch {
-      if (streamOpId) app.streamTracking.finish(streamOpId, 0, "Stream error");
-      sendBinaryError(reply, 502, "Stream error");
-      return;
-    }
-  } else {
-    const peer = app.peerRegistry.peers.get(best.instance_id);
-    if (!peer) {
-      if (streamOpId) app.streamTracking.finish(streamOpId, 0, "Peer not available");
-      sendBinaryError(reply, 502, "Peer not available");
-      return;
-    }
-    try {
-      const qs = streamParams.toString();
-      const path = `/federation/stream/${encodeURIComponent(best.remote_id)}${qs ? `?${qs}` : ""}`;
-      response = await app.federatedFetch(
-        peer,
-        path,
-        {
-          asUser: request.subsonicUser.username,
-          headers: rangeHeader ? { range: rangeHeader } : undefined,
-        },
-      );
-    } catch {
-      if (streamOpId) app.streamTracking.finish(streamOpId, 0, "Peer stream error");
-      sendBinaryError(reply, 502, "Peer stream error");
-      return;
-    }
-  }
-
-  if (!response.body) {
-    if (streamOpId) app.streamTracking.finish(streamOpId, 0, "Empty response from upstream");
-    sendBinaryError(reply, 502, "Empty response from upstream");
-    return;
-  }
-
-  const headers: Record<string, string> = {
-    "content-type": response.headers.get("content-type") || "audio/mpeg",
-  };
-  const contentLength = response.headers.get("content-length");
-  if (contentLength) headers["content-length"] = contentLength;
-  // `dlna=1` flag (set by `buildStreamUrl` for DLNA renderers, #218):
-  // strict renderers (WMP, Xbox) require `transferMode.dlna.org` +
-  // `contentFeatures.dlna.org` on the byte response, and tolerate a
-  // forced `accept-ranges: bytes` even when upstream omits it.
-  const isDlna = q.dlna === "1";
-  const acceptRanges =
-    response.headers.get("accept-ranges") || (isDlna ? "bytes" : null);
-  if (acceptRanges) headers["accept-ranges"] = acceptRanges;
-  if (isDlna) {
-    const reqHeaders = request.headers as Record<string, string | undefined>;
-    headers["transferMode.dlna.org"] =
-      reqHeaders["transfermode.dlna.org"] ?? "Streaming";
-    headers["contentFeatures.dlna.org"] =
-      "DLNA.ORG_OP=01;DLNA.ORG_FLAGS=01700000000000000000000000000000";
-  }
-  const contentRange = response.headers.get("content-range");
-  if (contentRange) headers["content-range"] = contentRange;
-
-  reply.raw.writeHead(response.status, headers);
-  const nodeStream = Readable.fromWeb(
-    response.body as import("node:stream/web").ReadableStream,
-  );
-  
-  // Track bytes transferred
-  nodeStream.on("data", (chunk) => {
-    bytesTransferred += chunk.length;
-    if (streamOpId) app.streamTracking.updateBytes(streamOpId, bytesTransferred);
-  });
-
-  // Finish tracking when stream ends or errors
-  nodeStream.on("end", () => {
-    if (streamOpId) {
-      app.streamTracking.finish(streamOpId, bytesTransferred, null);
-    }
-  });
-
-  nodeStream.on("error", (err) => {
-    if (streamOpId) {
-      app.streamTracking.finish(streamOpId, bytesTransferred, err instanceof Error ? err.message : String(err));
-    }
-  });
-  
-  nodeStream.pipe(reply.raw);
-}
-
-  binaryRoute("/stream", handleStream);
-  binaryRoute("/download", handleStream); // alias — clients use interchangeably
-
-  // ── star / unstar / getStarred / getStarred2 (#104) ─────────────────────────
-  //
-  // Per-user favorites, stored in the `user_stars` table on this hub. Targets
-  // are unified_*.id UUIDs; orphans (target gone after a sync) are dropped at
-  // read time via JOIN. Stars are local to the hub the user logs into and are
-  // not federated.
-
-  type StarKind = "track" | "album" | "artist";
-
-  function asArray(v: unknown): string[] {
-    if (v == null) return [];
-    return Array.isArray(v) ? (v as string[]) : [String(v)];
-  }
-
-  // Raw IDs are produced by `generateDeterministicId` (UUID v4 shape, all
-  // lowercase hex). Anchoring the classifier on this shape rejects malformed
-  // input (e.g. `id=tomato`) instead of silently inserting `omato` as a
-  // track target. Bare-UUID forms (no prefix) on `albumId`/`artistId` are
-  // also accepted, matching how some Subsonic clients send them.
-  const UUID_RE = /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/;
-
-  function classifyStarId(encoded: string): { kind: StarKind; raw: string } | null {
-    // Order matters: "ar"/"al" both start with "a"; "t" is a one-char prefix.
-    let kind: StarKind | null = null;
-    let raw = "";
-    if (encoded.startsWith("al")) {
-      kind = "album";
-      raw = encoded.slice(2);
-    } else if (encoded.startsWith("ar")) {
-      kind = "artist";
-      raw = encoded.slice(2);
-    } else if (encoded.startsWith("t")) {
-      kind = "track";
-      raw = encoded.slice(1);
-    }
-    if (!kind || !UUID_RE.test(raw)) return null;
-    return { kind, raw };
-  }
-
-  function unwrapKindId(encoded: string, prefix: string): string | null {
-    // Accept either the prefixed form (`al<uuid>`/`ar<uuid>`) or a bare UUID
-    // — the second is what some legacy clients send for `albumId`/`artistId`.
-    const raw = encoded.startsWith(prefix) ? encoded.slice(prefix.length) : encoded;
-    return UUID_RE.test(raw) ? raw : null;
-  }
-
-  function collectStarTargets(
-    q: Record<string, string | string[] | undefined>,
-  ): Array<{ kind: StarKind; raw: string }> {
-    const out: Array<{ kind: StarKind; raw: string }> = [];
-    for (const id of asArray(q.id)) {
-      const c = classifyStarId(id);
-      if (c) out.push(c);
-    }
-    for (const id of asArray(q.albumId)) {
-      const raw = unwrapKindId(id, "al");
-      if (raw) out.push({ kind: "album", raw });
-    }
-    for (const id of asArray(q.artistId)) {
-      const raw = unwrapKindId(id, "ar");
-      if (raw) out.push({ kind: "artist", raw });
-    }
-    return out;
-  }
-
-  route("/star", async (request, reply) => {
-    const q = request.query as Record<string, string | string[] | undefined>;
-    const targets = collectStarTargets(q);
-    starInsertTx(request.subsonicUser.id, targets);
-    sendSubsonicOk(reply, q as Record<string, string>, {});
-  });
-
-  route("/unstar", async (request, reply) => {
-    const q = request.query as Record<string, string | string[] | undefined>;
-    const targets = collectStarTargets(q);
-    starDeleteTx(request.subsonicUser.id, targets);
-    sendSubsonicOk(reply, q as Record<string, string>, {});
-  });
-
-  function buildStarredEnvelope(userId: string) {
-    const artists = starredArtistsStmt.all(userId) as Array<
-      ArtistRow & { starred_at: string }
-    >;
-    const albums = starredAlbumsStmt.all(userId) as Array<
-      ReleaseGroupRow & { starred_at: string }
-    >;
-    const songs = starredSongsStmt.all(userId) as Array<
-      TrackRow & { starred_at: string }
-    >;
-
-    const album = albums.map((a) => ({
-      ...buildAlbum(a),
-      starred: sqliteToIso(a.starred_at),
-    }));
-    const song = songs.map((s) => ({
-      ...buildSong(s),
-      starred: sqliteToIso(s.starred_at),
-    }));
-    annotatePlays(app.playEvents, userId, "album", "al", album);
-    annotatePlays(app.playEvents, userId, "track", "t", song);
-
-    return {
-      artist: artists.map((a) => ({
-        id: encodeId("ar", a.id),
-        name: a.name,
-        albumCount: a.albumCount,
-        coverArt: a.image_url ?? undefined,
-        starred: sqliteToIso(a.starred_at),
-      })),
-      album,
-      song,
-    };
-  }
-
-  route("/getStarred2", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-    const env = buildStarredEnvelope(request.subsonicUser.id);
-    sendSubsonicOk(reply, q, { starred2: env });
-  });
-
-  route("/getStarred", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-    const env = buildStarredEnvelope(request.subsonicUser.id);
-    sendSubsonicOk(reply, q, { starred: env });
-  });
-
-  // ── Playlist stubs ──────────────────────────────────────────────────────────
-  // TODO: implement fully once playlists table is populated (Phase 3+)
-
-  route("/getPlaylists", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-    sendSubsonicOk(reply, q, { playlists: { playlist: [] } });
-  });
-
-  route("/getPlaylist", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-    sendSubsonicError(reply, 70, "Playlist not found", q);
-  });
-
-  route("/createPlaylist", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-    sendSubsonicError(reply, 0, "Generic error: not yet implemented", q);
-  });
-
-  route("/updatePlaylist", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-    sendSubsonicError(reply, 0, "Generic error: not yet implemented", q);
-  });
-
-  route("/deletePlaylist", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-    sendSubsonicError(reply, 0, "Generic error: not yet implemented", q);
-  });
-
-  // ── Scrobble (#197) ───────────────────────────────────────────────────────
-  // The single play-recording path. Every playback surface reports its own
-  // play here: the SPA scrobbles for both local <audio> and Sonos-cast playback
-  // (off the actual playback position), and 3rd-party Subsonic apps scrobble
-  // too. `StreamTrackingService.finish()` records nothing — there is no
-  // server-side play inference — so this endpoint is the only writer of
-  // play_events. `submission=true` (the default) records a play; `false` is a
-  // "now playing" notification and is not counted. Accepts one or many `id`;
-  // unknown / malformed ids are skipped so a batch with one bad id still
-  // records the rest (Subsonic leniency). Honors the optional `time` param
-  // (epoch ms of when the play occurred) so offline/batching clients can
-  // backfill; absent, the play is stamped now.
-  route("/scrobble", async (request, reply) => {
-    const q = request.query as Record<string, string | string[] | undefined>;
-    const submission = q.submission == null ? true : String(q.submission) !== "false";
-    // Subsonic `time` is epoch milliseconds; ignore a non-numeric value.
-    const timeRaw = Array.isArray(q.time) ? q.time[0] : q.time;
-    const playedAtMs =
-      timeRaw != null && timeRaw !== "" && Number.isFinite(Number(timeRaw))
-        ? Number(timeRaw)
-        : null;
-
-    if (submission) {
-      const ids = (Array.isArray(q.id) ? q.id : q.id != null ? [q.id] : []).map(
-        String,
-      );
-      for (const encoded of ids) {
-        let trackId: string;
-        try {
-          trackId = decodeId(encoded, "t");
-        } catch {
-          continue; // not a track id — skip, don't fail the batch
-        }
-        // Only record plays of tracks this hub actually knows about; ignore
-        // unknown ids so a stale/garbage id can't pollute the history.
-        const track = queries.trackExists.get(trackId);
-        if (!track) continue;
-        // Attribute to the source we'd stream from (preferred source), local or
-        // peer. Best-effort: a known track with no source row records as null.
-        const src = queries.sourceForScrobble.get(trackId) as
-          | { instance_id: string }
-          | undefined;
-        app.playEvents.record({
-          userId: request.subsonicUser.id,
-          unifiedTrackId: trackId,
-          sourceInstanceId: src?.instance_id ?? null,
-          clientName: typeof q.c === "string" ? q.c : null,
-          playedAtMs,
-        });
-      }
-    }
-
-    sendSubsonicOk(reply, q as Record<string, string>, {});
-  });
-
-  route("/getNowPlaying", async (request, reply) => {
-    const q = request.query as Record<string, string>;
-    sendSubsonicOk(reply, q, { nowPlaying: { entry: [] } });
-  });
-};
